requirements.txt
# Medic Agent Dependencies
# Python 3.11+

# Core dependencies
pyyaml>=6.0.1

# Async HTTP client (for SIEM integration)
aiohttp>=3.9.0

# Redis client (for Smith event bus)
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
flake8>=6.1.0

# Optional: Web interface (Phase 2+)
# fastapi>=0.104.0
# uvicorn>=0.24.0

# Optional: Database (Phase 4+)
# sqlalchemy>=2.0.0
# aiosqlite>=0.19.0

# Optional: Learning/Analytics (Phase 4+)
# pandas>=2.0.0
# scikit-learn>=1.3.0
